Module euler_library::cards

Playing card library for poker to rank hands.

Examples

use euler_library::cards as eu_cards;

let hand_cs = "AD AS JH JS 2C".chars().filter(|&x| x != ' ' && x != '\n').collect::<Vec<char>>();
let hand_loser = eu_cards::get_hand(5, hand_cs);
assert_eq!(hand_loser.show(), "[(Two, Clubs), (Jack, Spades), (Jack, Hearts), (Ace, Spades), (Ace, Diamonds)]");

let hand_cs = "3D 3H 3C 2S 2D".chars().filter(|&x| x != ' ' && x != '\n').collect::<Vec<char>>();
let hand_winner = eu_cards::get_hand(5, hand_cs);
assert_eq!(hand_winner.show(), "[(Two, Diamonds), (Two, Spades), (Three, Clubs), (Three, Hearts), (Three, Diamonds)]");

assert!(hand_winner.get_rank() > hand_loser.get_rank());

Structs

Card

Hand

Enums

Suit

Val

Functions

char_to_suit
Returns Suit enum from a character.

char_to_val
Returns card face Val enum from a character.

get_hand
Return a Hand of length cnt from a vector of characters.

show_grp
Returns a new line seperated String of hands for printing.