Module euler_library::common

Common functions used in solving Project Euler problems.

Examples

use euler_library::common as eu;

assert!(eu::is_palindrome(12321));
assert_eq!(eu::divisor_sum_list(10), [0, 0, 1, 1, 3, 1, 6, 1, 7, 4, 8]);

let res = [[1, 1], [1, 2], [1, 3], [2, 1], [2, 2], [2, 3], [3, 1], [3, 2], [3, 3]];
assert_eq!(eu::perms_with_reps(2, &vec![1, 2, 3]), res);

Functions

accumulate
Returns vector of the running total of xs.

cartesian_product
Returns the cartesian product of a Vec of Vec of T.

continued_fraction
Return (numerator, denominator) after evaluating continued_fraction.

divisor_sum
Returns the sum of the proper divisors of n (not including n).

divisor_sum_list
Returns a vector containing divisor_sum(i) for i from 0 to n.

factorial
Returns factorial of n for numbers less than 21.

from_bytes
Returns Result from attempting to parses a Vec of u8.

from_digits
Returns a usize number from vector xs of usize digits.

is_palindrome
Returns true if value v is a palindrome (reads the same backward or forward).

is_pandigital
Returns true if string is pandigital over digits start to n.

is_perm
Returns true if a and b are permutations of each other.

k_nested_recur
Returns k nested 'loops' from xs, ordered.

perms_with_reps
Returns permutations k chosen from xs, odered, repetition allowed.

perms_without_reps_recur
Returns permutations k chosen from xs, odered, no repetition.

phis
Returns vector phi values (also called totient) from 0 to d.

replicate
Returns and iterator of length n of repeated values of elt

sqrt_terms
Returns continued fraction form of sqrt of n.

sum_of_digits
Returns the sum of the digits of a string.

to_bytes
Returns n as a vector of bytes.

to_digits
Returns usize n as a vector of its usize digits.